ap_template/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
numpy==1.24.4
pandas==2.0.3
replicate==0.26.1
psycopg2-binary==2.9.9
Flask==3.0.2
pyyaml==6.0.1
requests==2.28.1
openai==1.30.1
